MoveRecognitionPrompt/requirements.txt
# 语步识别工具依赖
zhipuai>=2.0.0
gradio>=4.0.0
python-dotenv>=1.0.0

# Web 服务依赖 (用于 ModelScope 风格页面)
flask>=3.0.0
flask-cors>=4.0.0
